add _form partial for create edit templates

+++ src/Controller/PinsController.php
@@ -87,11 +87,14 @@ class PinsController extends AbstractController
 }
 
     /**
-     * @Route("/pins/{id<[0-9]+>}/delete", name="app_pins_delete", methods={"GET"})
+     * @Route("/pins/{id<[0-9]+>}/delete", name="app_pins_delete", methods={"POST"})
      */
-    public function delete(Pin $pin, EntityManagerInterface $em): Response {
-        $em->remove($pin);
-        $em->flush();
+    public function delete(Request $request, Pin $pin, EntityManagerInterface $em): Response {
+          
+        if($this->isCsrfTokenValid('pin_deletion_' . $pin->getId(), $request->request->get('csrf_token') )) { 
+            $em->remove($pin);
+            $em->flush();
+        }
 
         return $this->redirectToRoute('app_home');
     }
